Separate implementation of the symbol-dce pass.
